backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
langchain==0.1.16
openai==1.23.6
psycopg2==2.9.9
PyJWT==2.8.0
